Extract PluginBase abstract class for common operations
Plugin interface is now lean — just onConfig() and pluginName().
PluginBase provides default implementations for:
- listTenantLogs (reads from MemoryAppender)
- setLogLevel (changes logback level at runtime)
- info (JVM/OS metadata)
- diagnostics (heap stats, processors)
- shutdown (logs warning)
- processLog (logs payload)

Integrations extend PluginBase and only override CRM-specific methods:
  public class FinviPlugin extends PluginBase {
      public boolean onConfig(...) { ... }
      public List<Pool> listPools() { ... }
      public void onTelephonyResult(...) { ... }
  }
